requirements.txt
attrs==23.2.0
blinker==1.7.0
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.2
cssselect==1.2.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.6
jmespath==1.0.1
kaitaistruct==0.10
lxml==5.1.0
mysql-connector-python==8.3.0
outcome==1.3.0.post0
packaging==23.2
parsel==1.8.1
pyasn1==0.5.1
pycparser==2.21
pydivert==2.1.0
pyOpenSSL==24.0.0
pyparsing==3.1.1
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.31.0
selenium==4.17.2
selenium-wire==5.1.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
undetected-chromedriver==3.5.4
urllib3==2.2.0
w3lib==2.1.2
websockets==12.0
wsproto==1.2.0
zstandard==0.22.0
